requirements.txt
appdirs==1.4.4
beartype==0.17.0
bypy==1.8.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
CoLT5-attention==0.10.19
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
local-attention==1.9.0
MarkupSafe==2.1.5
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.2.0
piq==0.8.0
protobuf==4.25.2
psutil==5.9.8
ptflops==0.7.2.2
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.2
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
timm==0.9.12
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.2.0
wandb==0.16.3
